# gaggiuino-local-profiler/go/internal/library/handlers.go
package library

import (
	"errors"
	"io"
	"net/http"
	"os"

	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/auth"
	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/httputil"
	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/img"
	"github.com/mxkissnr/gaggiuino-local-profiler/go/internal/shots"
)

// This file registers the library router on Go 1.22+'s method-and-wildcard
// http.ServeMux, the same pattern shots/handlers.go uses — see that file's
// header comment for the registration-order caveat that doesn't apply here
// either (ServeMux always prefers the most specific literal pattern, so e.g.
// "/api/library/beans-info" always wins over "/api/library/bean/{id}"
// regardless of registration order).
//
// scan.go's barcode-lookup handler registers through the same mux but lives
// in its own file given its SSRF-guard-heavy shape.

const jsonBodyLimit = 16 * 1024 // 16kb default body limit.

// Handlers wires Repository (+ a shots.Repository for grinder wear stats,
// see service.go's ComputeGrinderWearStats doc comment) into net/http
// handlers.
type Handlers struct {
	repo            *Repository
	shotsRepo       *shots.Repository
	imageDir        string
	limiter         *rateLimiter
	onGrinderDelete func(grinderID int64) error

	// onBeanRestocked runs after a new bag has been added to a bean
	// successfully (newBag). wasEmpty reports whether the bean had no
	// remaining stock before the bag was added — see newBag. Set via
	// SetOnBeanRestocked by cmd/server, which uses it to drive the restock
	// achievement (#1286 R2). A callback rather than a direct import for the
	// same reason as onGrinderDelete: internal/achievements already imports
	// internal/library, so the wiring has to run this direction. A nil hook
	// (never wired, e.g. in this package's own unit tests) is a no-op.
	onBeanRestocked func(wasEmpty bool)
}

// SetOnGrinderDeleted wires the maintenance domain's cleanup of a deleted
// grinder's `grinder_{id}` maintenance-table row as a callback rather than a
// direct import: this package already gets imported BY internal/maintenance
// (for grinder existence checks in canonicalTask() and grinder names in
// getMaintenance()/getMaintenanceLog()), so importing internal/maintenance
// back from here would close a cycle. cmd/server calls this once at startup,
// after both packages' Handlers exist — see main.go. A nil hook (never wired,
// e.g. in this package's own unit tests) is a no-op: deleting a grinder
// leaves its `maintenance` row in place (harmless — getMaintenance only
// iterates the library's actual grinders) — see this file's deleteGrinder doc
// comment.
func (h *Handlers) SetOnGrinderDeleted(fn func(grinderID int64) error) {
	h.onGrinderDelete = fn
}

// SetOnBeanRestocked wires the side effect to run after a new bag has been
// added to a bean successfully (#1286 R2). cmd/server uses it to let the
// achievements service see a `bean-changed` event with reason `restock`,
// which unlocks the restock badge when the bean had no stock left. wasEmpty
// is computed before the write — see newBag. internal/achievements imports
// internal/library, so wiring this as a callback here avoids the import
// cycle a direct dependency would create. A nil hook is a no-op; the
// callback never changes the response — the new-bag request itself already
// succeeded.
func (h *Handlers) SetOnBeanRestocked(fn func(wasEmpty bool)) {
	h.onBeanRestocked = fn
}

func (h *Handlers) notifyBeanRestocked(wasEmpty bool) {
	if h.onBeanRestocked != nil {
		h.onBeanRestocked(wasEmpty)
	}
}

// NewHandlers builds Handlers around repo and shotsDB (the same *sql.DB
// cmd/server already opens once and shares across every domain package —
// see shots.NewRepository's own call site in main.go). Image uploads are
// written under DefaultImageDir; tests override the unexported imageDir
// field to point at a throwaway directory instead (see helpers_test.go).
func NewHandlers(repo *Repository, shotsRepo *shots.Repository) *Handlers {
	return &Handlers{repo: repo, shotsRepo: shotsRepo, imageDir: DefaultImageDir, limiter: newRateLimiter()}
}

// RegisterRoutes registers every /api/library/* route onto mux.
func (h *Handlers) RegisterRoutes(mux *http.ServeMux) {
	mux.HandleFunc("GET /api/library", h.getLibrary)
	mux.HandleFunc("GET /api/library/beans-info", h.getBeansInfo)

	mux.HandleFunc("POST /api/library/bean", h.createBean)
	mux.HandleFunc("PUT /api/library/bean/{id}", h.updateBean)
	mux.HandleFunc("POST /api/library/bean/{id}/new-bag", h.newBag)
	mux.HandleFunc("POST /api/library/bean/{id}/reorder-bags", h.reorderBags)
	mux.HandleFunc("POST /api/library/bean/{id}/freeze-portions", h.freezePortions)
	mux.HandleFunc("POST /api/library/bean/{id}/thaw-portion", h.thawPortion)
	mux.HandleFunc("POST /api/library/bean/{id}/adjust-frozen-portion", h.adjustFrozenPortion)
	mux.HandleFunc("PUT /api/library/bean/{id}/bag/{bagId}", h.updateBag)
	mux.HandleFunc("DELETE /api/library/bean/{id}/bag/{bagId}", h.deleteBag)
	mux.HandleFunc("POST /api/library/bean/{id}/delete", h.deleteBean)
	mux.HandleFunc("POST /api/library/bean/{id}/toggle-active", h.toggleBeanActive)
	mux.HandleFunc("POST /api/library/bean/{id}/known-grind", h.knownGrind)
	mux.HandleFunc("GET /api/library/bean/{id}/image", h.getBeanImage)
	mux.HandleFunc("POST /api/library/bean/{id}/image", h.postBeanImage)

	mux.HandleFunc("POST /api/library/grinder", h.createGrinder)
	mux.HandleFunc("PUT /api/library/grinder/{id}", h.updateGrinder)
	mux.HandleFunc("POST /api/library/grinder/{id}/reset-burrs", h.resetBurrs)
	mux.HandleFunc("PUT /api/library/grinder/{id}/zero-point", h.setGrinderZeroPoint)
	mux.HandleFunc("DELETE /api/library/grinder/{id}/zero-point/{since}", h.deleteGrinderZeroPoint)
	mux.HandleFunc("POST /api/library/grinder/{id}/delete", h.deleteGrinder)
	mux.HandleFunc("GET /api/library/grinder/{id}/image", h.getGrinderImage)
	mux.HandleFunc("POST /api/library/grinder/{id}/image", h.postGrinderImage)

	mux.HandleFunc("GET /api/library/baskets", h.listBaskets)
	mux.HandleFunc("POST /api/library/basket", h.createBasket)
	mux.HandleFunc("PUT /api/library/basket/{id}", h.updateBasket)
	mux.HandleFunc("DELETE /api/library/basket/{id}", h.deleteBasket)
	mux.HandleFunc("GET /api/library/basket/{id}/image", h.getBasketImage)
	mux.HandleFunc("POST /api/library/basket/{id}/image", h.postBasketImage)

	mux.HandleFunc("GET /api/library/puckscreens", h.listPuckScreens)
	mux.HandleFunc("POST /api/library/puckscreen", h.createPuckScreen)
	mux.HandleFunc("PUT /api/library/puckscreen/{id}", h.updatePuckScreen)
	mux.HandleFunc("DELETE /api/library/puckscreen/{id}", h.deletePuckScreen)
	mux.HandleFunc("GET /api/library/puckscreen/{id}/image", h.getPuckScreenImage)
	mux.HandleFunc("POST /api/library/puckscreen/{id}/image", h.postPuckScreenImage)

	mux.HandleFunc("GET /api/library/milks", h.listMilks)
	mux.HandleFunc("POST /api/library/milk", h.createMilk)
	mux.HandleFunc("PUT /api/library/milk/{id}", h.updateMilk)
	mux.HandleFunc("DELETE /api/library/milk/{id}", h.deleteMilk)
	mux.HandleFunc("POST /api/library/milk/{id}/deduct", h.deductMilk)
	mux.HandleFunc("POST /api/library/milk/{id}/restock", h.restockMilk)

	mux.HandleFunc("POST /api/library/recipe", h.createRecipe)
	mux.HandleFunc("PUT /api/library/recipe/{id}", h.updateRecipe)
	mux.HandleFunc("POST /api/library/recipe/{id}/delete", h.deleteRecipe)

	h.registerScanRoute(mux)
}

// ── response helpers (see internal/httputil) ─────────────────────────────

var (
	writeJSON  = httputil.WriteJSON
	writeError = httputil.WriteError
)

func internalError(w http.ResponseWriter, err error) {
	httputil.InternalError(w, "library", err)
}

// apiError carries one of this handler layer's existing 4xx outcomes out of a
// Repository.Update callback. Update runs its callback while holding the
// library write lock and an HTTP response must not be written under it, so a
// callback returns one of these instead and the handler writes the response
// once Update has returned: writeUpdateError maps it back onto the exact
// status/message the pre-Update code wrote directly.
type apiError struct {
	status  int
	message string
}

func (e *apiError) Error() string { return e.message }

var (
	errNotFound              = &apiError{http.StatusNotFound, "not found"}
	errNoActiveBag           = &apiError{http.StatusBadRequest, "no active bag"}
	errPortionFieldsRequired = &apiError{http.StatusBadRequest, "portionCount and portionWeight_g required"}
	errCannotDeleteLastBag   = &apiError{http.StatusBadRequest, "cannot delete last bag"}
	errFrozenPortionNotFound = &apiError{http.StatusNotFound, "frozen portion not found"}
	errInvalidPortionWeight  = &apiError{http.StatusBadRequest, "invalid portionWeight_g"}
	errInvalidFrozenAt       = &apiError{http.StatusBadRequest, "invalid frozenAt"}
	errInvalidRemainingCount = &apiError{http.StatusBadRequest, "invalid remainingCount"}
)

// writeUpdateError writes the response for an error returned by
// Repository.Update: a callback-classified *apiError becomes its recorded
// status/message, anything else the usual internalError.
func writeUpdateError(w http.ResponseWriter, err error) {
	var ae *apiError
	if errors.As(err, &ae) {
		writeError(w, ae.status, ae.message)
		return
	}
	internalError(w, err)
}

// entityExists reports whether id names an existing entity of one library
// kind. The *...Image handlers run this read-only check BEFORE reading or
// writing the uploaded file, so an unknown id still 404s — and no file is
// written — before any image validation, matching dev's original ordering.
// GetLibrary is always called first (even for an unparseable id), so a
// broken DB outranks the not-found response exactly as before; find is the
// kind's findXIndex.
func (h *Handlers) entityExists(id int64, noMatch bool, find func(Library, int64) int) (bool, error) {
	lib, err := h.repo.GetLibrary()
	if err != nil {
		return false, err
	}
	if noMatch {
		return false, nil
	}
	return find(lib, id) != -1, nil
}

func decodeJSONBody(w http.ResponseWriter, r *http.Request) (Entity, bool) {
	body, ok := httputil.DecodeJSONBody[Entity](w, r, jsonBodyLimit)
	if !ok {
		return nil, false
	}
	if body == nil {
		body = Entity{}
	}
	return body, true
}

// rateLimitCreate applies the `lib:<ip>`-keyed guard every library
// create-endpoint (bean/grinder/basket/milk/puckscreen/recipe) opens with.
func (h *Handlers) rateLimitCreate(w http.ResponseWriter, r *http.Request) bool {
	if h.limiter.allow("lib:"+auth.RemoteIP(r), 30) {
		return true
	}
	writeError(w, http.StatusTooManyRequests, "Rate limit exceeded")
	return false
}

// imageRateLimitPerMin is the dedicated per-IP ceiling for every library
// entity's POST .../image upload route (bean/grinder/basket/puckscreen,
// #1056). Each upload runs a full decode, downscale, re-encode and
// thumbnail pass (img.Save -> img.Optimize) on up to img.MaxBytes — the
// most CPU-expensive authenticated route family in the app without a
// feature limit before this. Set to the same ceiling rateLimitCreate
// already uses for library creates, and matches the shot-card renderer's
// own 30/min (internal/shots.cardRateLimitPerMin). Deliberately stricter:
// the shared 600/min backstop would otherwise be the only limit on image
// uploads, since creates/scans are the only routes with a feature limit.
const imageRateLimitPerMin = 30

// rateLimitImage applies the same `lib:<ip>`-shaped guard as rateLimitCreate,
// scoped to its own `image:<ip>` key so image uploads don't share a budget
// with entity creates.
func (h *Handlers) rateLimitImage(w http.ResponseWriter, r *http.Request) bool {
	if h.limiter.allow("image:"+auth.RemoteIP(r), imageRateLimitPerMin) {
		return true
	}
	writeError(w, http.StatusTooManyRequests, "Rate limit exceeded")
	return false
}

// ── GET /api/library, GET /api/library/beans-info ─────────────────────────

// getLibrary serves GET /api/library: the full Library object, grinders
// enriched with a computed `wear` field on read (not stored) — see
// service.go's ComputeGrinderWearStats.
func (h *Handlers) getLibrary(w http.ResponseWriter, r *http.Request) {
	lib, err := h.repo.GetLibrary()
	if err != nil {
		internalError(w, err)
		return
	}
	var allShots []shots.Shot
	if len(lib.Grinders) > 0 {
		allShots, err = h.shotsRepo.FindAllExcludingTrash()
		if err != nil {
			internalError(w, err)
			return
		}
	}
	grinders := make([]Entity, len(lib.Grinders))
	for i, g := range lib.Grinders {
		shotsSince, gramsSince := ComputeGrinderWearFrom(allShots, g)
		grinders[i] = withWearEntity(g, shotsSince, gramsSince)
	}
	lib.Grinders = grinders

	// Bags get the same read-time, non-persisted enrichment as grinders'
	// wear above — consumedG/remainingG/current per bag, computed by
	// SimulateBagQueue, so the frontend never replays doseRows itself (see
	// #sortOrder rework, decorateBeanStatus).
	if len(lib.Beans) > 0 {
		doseRows, err := h.shotsRepo.GetAnnotatedDoses()
		if err != nil {
			internalError(w, err)
			return
		}
		beans := make([]Entity, len(lib.Beans))
		for i, b := range lib.Beans {
			beans[i] = decorateBeanStatus(b, doseRows, lib.Beans)
		}
		lib.Beans = beans
	}
	writeJSON(w, http.StatusOK, lib)
}

// withWear attaches wear: computeGrinderWearStats(g). Field names are
// shotsSinceBurrs/gramsSinceBurrs, the actual return shape — NOT
// openapi.yaml's Grinder.wear schema (documented there as `{shots, grams}`),
// which drifted from the implementation; the running behavior wins, same
// "doc vs. code disagree, code wins" rule shots/doc.go states.
func (h *Handlers) withWear(grinder Entity) Entity {
	shotsSince, gramsSince, err := ComputeGrinderWearStats(h.shotsRepo, grinder)
	if err != nil {
		return withWearEntity(grinder, 0, 0)
	}
	return withWearEntity(grinder, shotsSince, gramsSince)
}

// withWearEntity returns a copy of grinder with the computed `wear` field
// attached — the enrichment step shared by the single-grinder withWear and
// getLibrary's single-pass loop.
func withWearEntity(grinder Entity, shotsSince int, gramsSince float64) Entity {
	out := make(Entity, len(grinder)+1)
	for k, v := range grinder {
		out[k] = v
	}
	out["wear"] = Entity{"shotsSinceBurrs": shotsSince, "gramsSinceBurrs": gramsSince}
	return out
}

// getBeansInfo serves GET /api/library/beans-info.
func (h *Handlers) getBeansInfo(w http.ResponseWriter, r *http.Request) {
	lib, err := h.repo.GetLibrary()
	if err != nil {
		internalError(w, err)
		return
	}
	writeJSON(w, http.StatusOK, GetBeansInfo(lib))
}

// ── image handling shared by bean/grinder/basket/puckscreen ───────────────

// serveImage handles the repeated `GET .../image` shape: 404 with
// {error:"no image"} when the entity/image is missing, otherwise serves the
// file with a 24h cache header.
func (h *Handlers) serveImage(w http.ResponseWriter, r *http.Request, ext, prefix string, id int64) {
	contentType, known := img.ExtContentType[ext]
	if ext == "" || !known {
		writeError(w, http.StatusNotFound, "no image")
		return
	}
	// `?thumb=1` serves the downscaled variant when present and otherwise
	// falls back to the full image — a missing thumbnail never 404s.
	thumb := r.URL.Query().Get("thumb") == "1"
	path := img.ServePath(h.imageDir, id, ext, prefix, thumb)
	if _, err := os.Stat(path); err != nil {
		writeError(w, http.StatusNotFound, "no image")
		return
	}
	w.Header().Set("Cache-Control", "public, max-age=86400")
	w.Header().Set("Content-Type", contentType)
	http.ServeFile(w, r, path)
}

// readUploadedImage handles the repeated raw-upload body handling every
// entity's `POST .../image` route uses: an unrecognized Content-Type or an
// empty body is "no image data" (400); an oversized body is rejected by
// MaxBytesReader before ever reaching img.Save's own size check.
func readUploadedImage(w http.ResponseWriter, r *http.Request) (data []byte, contentType string, ok bool) {
	contentType = r.Header.Get("Content-Type")
	_, typeKnown := img.ContentTypeKnown(contentType)
	r.Body = http.MaxBytesReader(w, r.Body, img.MaxBytes)
	data, err := io.ReadAll(r.Body)
	if err != nil {
		writeError(w, http.StatusRequestEntityTooLarge, "request entity too large")
		return nil, "", false
	}
	if !typeKnown || len(data) == 0 {
		writeError(w, http.StatusBadRequest, "no image data")
		return nil, "", false
	}
	return data, contentType, true
}