# gaggiuino-local-profiler/go/internal/system/openapi.yaml
openapi: 3.0.3
info:
  title: GLP REST API
  version: 2.0.0
  description: >
    REST API for the **Gaggiuino Local Profiler** Home Assistant add-on.
    Provides shot storage, coffee library, maintenance tracking, a barista
    orders system, and live machine data.
  contact:
    url: https://github.com/mxkissnr/gaggiuino-local-profiler
servers:
  - url: http://localhost:8099
    description: Direct port (integration, cards)
  - url: /api/hassio_ingress/gaggiuino_local_profiler
    description: HA Ingress (browser — auth via HA session, no token needed)
security:
  - tokenAuth: []
components:
  securitySchemes:
    tokenAuth:
      type: apiKey
      in: header
      name: X-GLP-Token
      description: >
        64-char hex token, auto-generated on first start, stored in
        `/data/api_token.txt`. Retrieve from `GET /api/status`.
        Requests via HA Ingress bypass this check (HA session is sufficient).
        `GET /api/status` is always public.
    tokenQueryAuth:
      type: apiKey
      in: query
      name: token
      description: >
        Same token as `tokenAuth`, accepted as a `?token=` query parameter
        instead of the `X-GLP-Token` header. Only honored on `GET
        /api/events`, since a browser-native EventSource connection can't
        set custom headers — every other route requires the header.
  schemas:
    Ok:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean, example: true }
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: "not found" }
...
    NewBagInput:
      type: object
      properties:
        roastDate:   { type: string, format: date }
        stock_g:     { type: number, nullable: true }
        batchNumber: { type: string, maxLength: 50 }
...
paths:
...
  /api/library/bean:
    post:
      tags: [Library]
      summary: Add a bean to the library
      description: >
        Accepts the full Bean field set (see the Bean schema) — name is the
        only required field; `id` is server-assigned and ignored if sent.
        Also fires a fire-and-forget geocode (if `region` given) and/or image
        download (if `imageUrl` given, not a stored field).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Bean' }
      responses:
        "200": { description: Created bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "429": { description: Rate limit exceeded }
  /api/library/bean/{id}:
    put:
      tags: [Library]
      summary: Update a bean (partial — omitted fields keep their current value)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Bean' }
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/library/bean/{id}/new-bag:
    post:
      tags: [Library]
      summary: Open a new bag for a bean (joins the back of the queue)
      description: The new bag's roastDate/stock_g only become the bean's own display fields when it's also the bag SimulateBagQueue resolves as current (i.e. every other bag is already exhausted) — adding a bag while the current one still has stock does not disturb the bean-level fields.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewBagInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "404": { description: Not found }
  /api/library/bean/{id}/reorder-bags:
    post:
      tags: [Library]
      summary: Bulk-reassign sortOrder for a bean's "upcoming" (non-current) bags
      description: bagIds must list every upcoming bag exactly once (no duplicates, none omitted) — the server assigns sequential sortOrder values in that order, all strictly above the current bag's own position.
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderBagsInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: bagIds missing, containing a duplicate/unknown/current-bag id, or not covering every upcoming bag, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "404": { description: Bean not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/library/bean/{id}/freeze-portions:
    post:
      tags: [Library]
      summary: Freeze a portion of the active bag (#472) — pauses that portion's freshness clock
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FreezePortionsInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: No active bag, or portionCount/portionWeight_g missing/invalid }
        "404": { description: Not found }
  /api/library/bean/{id}/thaw-portion:
    post:
      tags: [Library]
      summary: Thaw one or more remaining portions of a frozen-portion batch
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ThawPortionInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "404": { description: Bean or frozen portion not found }
  /api/library/bean/{id}/adjust-frozen-portion:
    post:
      tags: [Library]
      summary: Correct a frozen-portion entry's weight/date/remaining count after the fact (#472)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustFrozenPortionInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: Invalid portionWeight_g/frozenAt/remainingCount }
        "404": { description: Bean or frozen portion not found }
  /api/library/bean/{id}/bag/{bagId}:
    put:
      tags: [Library]
      summary: Full-replace one bag's editable fields (roastDate/stock_g/price_eur/batchNumber; sortOrder optional)
      description: roastDate/stock_g/price_eur/batchNumber are full-replaced — resend the bag's current value for any field not being changed, or it's cleared. sortOrder is the one exception, falling back to the bag's existing value when omitted (reorder-bags is the normal bulk way to change it).
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: bagId, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BagUpdateInput'
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: Invalid roastDate/stock_g/price_eur, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "404": { description: Bean or bag not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
    delete:
      tags: [Library]
      summary: Delete a bag from a bean's bag history (the last remaining bag cannot be deleted)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: bagId, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: Cannot delete the last bag }
        "404": { description: Not found }
  /api/library/bean/{id}/delete:
    post:
      tags: [Library]
      summary: Delete a bean (and its photo, if any)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/library/bean/{id}/toggle-active:
    post:
      tags: [Library]
      summary: Toggle a bean's `enabled` flag (order card's bean picker override, independent of stock, #578)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "404": { description: Not found }
  /api/library/bean/{id}/known-grind:
    post:
      tags: [Library]
      summary: Remember the winning (grinder, grindSetting) combo for a bean (Guided Dial-In, #310)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grinder, grindSetting]
              properties:
                grinder:      { type: string, maxLength: 200 }
                grindSetting: { type: string, maxLength: 50 }
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: grinder or grindSetting missing }
        "404": { description: Not found }
  /api/library/bean/{id}/image:
    get:
      tags: [Library]
      summary: Serve a bean's photo (auto-imported or uploaded)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Image bytes, content: { image/*: { schema: { type: string, format: binary } } } }
        "404": { description: No image for this bean }
    post:
      tags: [Library]
      summary: Upload a bean photo manually (raw bytes, Content-Type selects the extension)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png:  { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Updated bean, content: { application/json: { schema: { $ref: '#/components/schemas/Bean' } } } }
        "400": { description: No image data, or unsupported image type }
        "404": { description: Not found }
  /api/library/grinder:
    post:
      tags: [Library]
      summary: Add a grinder to the library
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:         { type: string, maxLength: 200 }
                notes:        { type: string, maxLength: 1000 }
                burrType:     { type: string, maxLength: 200 }
                purchaseDate: { type: string, format: date }
                burrsResetAt: { type: string, format: date, description: "Defaults to purchaseDate if omitted" }
      responses:
        "200": { description: Created grinder, content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "400": { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "429": { description: Rate limit exceeded }
  /api/library/grinder/{id}:
    put:
      tags: [Library]
      summary: Update a grinder
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:         { type: string, maxLength: 200 }
                notes:        { type: string, maxLength: 1000 }
                burrType:     { type: string, maxLength: 200 }
                purchaseDate: { type: string, format: date }
                burrsResetAt: { type: string, format: date }
      responses:
        "200": { description: Updated grinder, content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/library/grinder/{id}/reset-burrs:
    post:
      tags: [Library]
      summary: Mark burrs as replaced — resets the wear counter to zero (calendar maintenance untouched)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: "Updated grinder (with recomputed `wear`)", content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "404": { description: Not found }
  /api/library/grinder/{id}/zero-point:
    put:
      tags: [Library]
      summary: "Log a new zero-point activation — corrects grind-setting suggestions/comparisons for drift (e.g. after cleaning) without rewriting any past shot's recorded value"
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [zeroPoint]
              properties:
                zeroPoint: { type: number }
                since: { type: integer, description: "Unix ms epoch. Omit or 0 to log the entry at the current time; set to insert a retroactive entry (e.g. a burr change that happened in the past)." }
      responses:
        "200": { description: "Updated grinder (with the new zeroPointHistory entry)", content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "400": { description: Invalid zeroPoint, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/library/grinder/{id}/zero-point/{since}:
    delete:
      tags: [Library]
      summary: "Remove one zero-point history entry (identified by its `since` unix-ms timestamp). Idempotent — succeeds even if no matching entry exists."
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: since, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: "Updated grinder (entry removed)", content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "404": { description: Not found, content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } } }
  /api/library/grinder/{id}/delete:
    post:
      tags: [Library]
      summary: Delete a grinder (also removes its photo and maintenance entry)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200": { description: Success, content: { application/json: { schema: { $ref: '#/components/schemas/Ok' } } } }
  /api/library/grinder/{id}/image:
    get:
      tags: [Library]
      summary: Serve a grinder's photo
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Image bytes, content: { image/*: { schema: { type: string, format: binary } } } }
        "404": { description: No image for this grinder }
    post:
      tags: [Library]
      summary: Upload a grinder photo (raw bytes, Content-Type selects the extension)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png:  { schema: { type: string, format: binary } }
          image/webp: { schema: { type: string, format: binary } }
      responses:
        "200": { description: Updated grinder, content: { application/json: { schema: { $ref: '#/components/schemas/Grinder' } } } }
        "400": { description: No image data, or unsupported image type }
        "404": { description: Not found }
...
  /api/library/basket:
    post:
      tags: [Library]
      summary: Add a basket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:         { type: string, maxLength: 200 }
                doseCapacity: { type: string, maxLength: 50 }
                wallType:     { type: string, enum: [pressurized, single-wall, precision-machined, high-flow] }
                shape:        { type: string, enum: [straight, tapered] }
                holeCount:    { type: string, maxLength: 50 }
                notes:        { type: string, maxLength: 1000 }
      responses:
        "200": { description: Created basket, content: { application/json: { schema: { $ref: '#/components/schemas/Basket' } } } }
        "400": { description: "name required, or invalid wallType/shape" }
        "429": { description: Rate limit exceeded }
...
  /api/library/milk:
    post:
      tags: [Library]
      summary: Add a milk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:    { type: string, maxLength: 100 }
                emoji:   { type: string, default: "🥛" }
                stockMl: { type: number, default: 0 }
      responses:
        "200": { description: Created milk, content: { application/json: { schema: { $ref: '#/components/schemas/Milk' } } } }
        "400": { description: name required }
        "429": { description: Rate limit exceeded }
...